Refactor: Move error tracking to diagnostics module
