+++ packages/analytics-core/src/diagnostics/diagnostics-uncaught-sdk-error.ts
